refactor(series): extract metadata hook, generic content filtering, detail panel
Series.tsx drops from 2,167 to 1,707 lines (component logic ~1,092 ->
~635; the rest is the in-page CSS string, kept for identical styling):

- useSeriesMetadata: TMDB series search + episode-title resolution
  with its episode cache, preserved verbatim.
- useContentFiltering<T>: kids-profile + parental gating extracted
  GENERIC over movies/series (IPC channel, indexedDB cache APIs and
  TMDB search dispatch on contentType) — VOD.tsx has a near-duplicate
  that can migrate next with zero string changes; untouched for now.
- SeriesDetailPanel: season tabs/episode selector block (currently
  display:none in favor of ContentDetailModal — preserved as is).

Grid (freshly virtualized), resume modal, player orchestration and
all service calls stay in the page, unchanged.
